Fix audit issue: Non-Decrementing Whitelist
CUZTokenSale
Within the ​ buyTokens​ function, there is a check that the purchase amount does not go over the
value listed in the whitelist. This check works as intended, however, since the amount listed in
the whitelist is not updated with the proper permitted amount, the contract allows repeated
purchases for the full original whitelisted amount.