Fix audit issue: Unable to Transfer Ownership
CUZTokenSale
The TokenSale contract is required to be made the owner of the token contract, in order to mint
the tokens. Because of this, the contracts are unable to transfer ownership of the token after it is
transferred to the sale instance, and also lose the ability to disable minting.