Fix audit issue: Possible Contract Lock
CUZFutureDevelopmentWallet
The release function contained in this contract contains a require that compares the balance,
minus the paid balance, and a constant. As this equation requires an exact token balance, if a
single token is deposited, as opposed to the specific value needed, this contract will be unable to
transfer funds via its release system.